refactor: consolidate array, null, and scalar processing into BuiltInTypeProcessor
- Removed ArrayProcessor, NullProcessor, and ScalarProcessor.
- Introduced BuiltInTypeProcessor to handle null, scalar, and array types efficiently.
- Updated ContextProcessorManager to register BuiltInTypeProcessor.
- Added unit tests for BuiltInTypeProcessor to ensure correct functionality.
